lingdoc-ai-service/requirements.txt
# =============================================================================
# LingDoc AI Service - Python Dependencies
# =============================================================================
# 说明：本文件定义 lingdoc-ai-service 的所有 Python 依赖包
# 使用方式：pip install -r requirements.txt
# =============================================================================

# --- Web Framework ---
# FastAPI: 高性能异步 Web 框架，用于构建 RESTful API
fastapi==0.115.0

# Uvicorn: ASGI 服务器，运行 FastAPI 应用
# 使用方式：uvicorn app.main:app --reload
uvicorn[standard]==0.32.0

# Python 多部分表单解析（FastAPI 文件上传依赖）
python-multipart==0.0.17

# --- OCR Engine ---
# PaddleOCR: 百度开源 OCR 引擎，支持中文识别
# 注意：首次运行会自动下载模型（约 100MB）
paddleocr==2.8.1

# PaddlePaddle: PaddleOCR 底层深度学习框架（CPU 版本）
paddlepaddle==2.6.1

# --- Document Processing ---
# PyMuPDF (fitz): PDF 文件处理（提取图片、文本）
pymupdf==1.24.14

# python-docx: Word 文档处理（.docx 格式）
python-docx==1.1.2

# Pillow: 图像处理（OCR 前置的图片转换）
Pillow==10.4.0

# --- LLM Client (阿里云 DashScope) ---
# dashscope: 阿里云大模型服务 SDK，支持 Qwen3-Max
# 文档：https://help.aliyun.com/document_detail/611473.html
dashscope==1.20.14

# --- HTTP Client ---
# httpx: 现代异步 HTTP 客户端（比 requests 更快）
httpx==0.27.2

# --- Data Validation ---
# pydantic: 数据模型验证（FastAPI 已内置，单独列出方便版本锁定）
pydantic==2.9.2

# --- Utilities ---
# python-dotenv: 从 .env 文件加载环境变量
python-dotenv==1.0.1

# aiosqlite: 异步 SQLite 驱动（如需要访问 Vault 数据库）
aiosqlite==0.20.0

# --- Testing ---
# pytest: 单元测试框架
pytest==8.3.3
pytest-asyncio==0.24.0

# --- Development (可选) ---
# black: 代码格式化
# black==24.10.0
